Add files via upload
We are going to obtain users emotion at the moment. It uses tkinter as a sketchy
demonstration of user interface.Users are able to select multiple emotions.
There are 10 categories of emotion the system presented to users to choose from. These are 5 postive emotions ("Happy", "Satisfied", "Peaceful", "Excited", "Content") and 5 negative emotions ("Sad", "Angry", "Fearful", "Depressed", "Sorrowful"). These emotions taken as inputs from the GUI interface we built through tkinter.
The correspondence of every emotion with genre of movies are set up as below:
• Happy - Horror
• Sad - Drama
• Satisfied - Animation
• Angry - Romance
• Peaceful - Fantasy
• Fearful - Adventure
• Excited - Crime
• Depressed - Comedy
• Content - Mystery
• Sorrowful - Action
Based on the inputted emotion, the system is going to be selected from the corresponding genre based on their ratings given by two websites: IMDB and Rotten Tomatoes. The reason why we are collecting movie information from both websites is that we believe the system is able to capture a more full-scaled opinions from movie lovers.